gmsol_store

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AcceptReceiver

Generated client accounts for AcceptReceiver.

AcceptReferralCode

Generated client accounts for AcceptReferralCode.

AcceptStoreAuthority

Generated client accounts for AcceptStoreAuthority.

CancelOrderIfNoPosition

Generated client accounts for CancelOrderIfNoPosition.

CancelReferralCodeTransfer

Generated client accounts for CancelReferralCodeTransfer.

CheckRole

Generated client accounts for CheckRole.

ClaimFeesFromMarket

Generated client accounts for ClaimFeesFromMarket.

ClearAllPrices

Generated client accounts for ClearAllPrices.

CloseDeposit

Generated client accounts for CloseDeposit.

CloseEmptyClaimableAccount

Generated client accounts for CloseEmptyClaimableAccount.

CloseGlvDeposit

Generated client accounts for CloseGlvDeposit.

CloseGlvShift

Generated client accounts for CloseGlvShift.

CloseGlvWithdrawal

Generated client accounts for CloseGlvWithdrawal.

CloseGtExchange

Generated client accounts for CloseGtExchange.

CloseMarketConfigBuffer

Generated client accounts for CloseMarketConfigBuffer.

CloseOrder

Generated client accounts for CloseOrder.

CloseShift

Generated client accounts for CloseShift.

CloseWithdrawal

Generated client accounts for CloseWithdrawal.

ConfigurateGt

Generated client accounts for ConfigurateGt.

ConfirmGtExchangeVault

Generated client accounts for ConfirmGtExchangeVault.

CreateDeposit

Generated client accounts for CreateDeposit.

CreateGlvDeposit

Generated client accounts for CreateGlvDeposit.

CreateGlvShift

Generated client accounts for CreateGlvShift.

CreateGlvWithdrawal

Generated client accounts for CreateGlvWithdrawal.

CreateOrder

Generated client accounts for CreateOrder.

CreateShift

Generated client accounts for CreateShift.

CreateWithdrawal

Generated client accounts for CreateWithdrawal.

DisableRole

Generated client accounts for DisableRole.

EnableRole

Generated client accounts for EnableRole.

ExecuteDecreaseOrder

Generated client accounts for ExecuteDecreaseOrder.

ExecuteDeposit

Generated client accounts for ExecuteDeposit.

ExecuteGlvDeposit

Generated client accounts for ExecuteGlvDeposit.

ExecuteGlvShift

Generated client accounts for ExecuteGlvShift.

ExecuteGlvWithdrawal

Generated client accounts for ExecuteGlvWithdrawal.

ExecuteIncreaseOrSwapOrder

Generated client accounts for ExecuteIncreaseOrSwapOrder.

ExecuteShift

Generated client accounts for ExecuteShift.

ExecuteWithdrawal

Generated client accounts for ExecuteWithdrawal.

GrantRole

Generated client accounts for GrantRole.

HasRole

Generated client accounts for HasRole.

Initialize

Generated client accounts for Initialize.

InitializeGlv

Generated client accounts for InitializeGlv.

InitializeGt

Generated client accounts for InitializeGt.

InitializeMarket

Generated client accounts for InitializeMarket.

InitializeMarketConfigBuffer

Generated client accounts for InitializeMarketConfigBuffer.

InitializeMarketVault

Generated client accounts for InitializeMarketVault.

InitializeOracle

Generated client accounts for InitializeOracle.

InitializePriceFeed

Generated client accounts for InitializePriceFeed.

InitializeReferralCode

Generated client accounts for InitializeReferralCode.

InitializeTokenMap

Generated client accounts for InitializeTokenMap.

InsertConfig

Generated client accounts for InsertConfig.

InsertGlvMarket

Generated client accounts for InsertGlvMarket.

MarketTransferIn

Generated client accounts for MarketTransferIn.

MigrateReferralCode

Generated client accounts for MigrateReferralCode.

PositionCut

Generated client accounts for PositionCut.

PrepareAssociatedTokenAccount

Generated client accounts for PrepareAssociatedTokenAccount.

PrepareGtExchangeVault

Generated client accounts for PrepareGtExchangeVault.

PreparePosition

Generated client accounts for PreparePosition.

PrepareTradeEventBuffer

Generated client accounts for PrepareTradeEventBuffer.

PrepareUser

Generated client accounts for PrepareUser.

PushToMarketConfigBuffer

Generated client accounts for PushToMarketConfigBuffer.

PushToTokenMap

Generated client accounts for PushToTokenMap.

PushToTokenMapSynthetic

Generated client accounts for PushToTokenMapSynthetic.

ReadMarket

Generated client accounts for ReadMarket.

ReadMarketWithToken

Generated client accounts for ReadMarketWithToken.

ReadTokenMap

Generated client accounts for ReadTokenMap.

RemoveGlvMarket

Generated client accounts for RemoveGlvMarket.

RequestGtExchange

Generated client accounts for RequestGtExchange.

RevokeRole

Generated client accounts for RevokeRole.

SetExpectedProvider

Generated client accounts for SetExpectedProvider.

SetFeedConfig

Generated client accounts for SetFeedConfig.

SetMarketConfigBufferAuthority

Generated client accounts for SetMarketConfigBufferAuthority.

SetPricesFromPriceFeed

Generated client accounts for SetPricesFromPriceFeed.

SetReferrer

Generated client accounts for SetReferrer.

SetTokenMap

Generated client accounts for SetTokenMap.

ToggleFeature

Generated client accounts for ToggleFeature.

ToggleGTMinting

Generated client accounts for ToggleGTMinting.

ToggleMarket

Generated client accounts for ToggleMarket.

ToggleTokenConfig

Generated client accounts for ToggleTokenConfig.

TransferReceiver

Generated client accounts for TransferReceiver.

TransferReferralCode

Generated client accounts for TransferReferralCode.

TransferStoreAuthority

Generated client accounts for TransferStoreAuthority.

UpdateAdlState

Generated client accounts for UpdateAdlState.

UpdateGlvConfig

Generated client accounts for UpdateGlvConfig.

UpdateGlvMarketConfig

Generated client accounts for UpdateGlvMarketConfig.

UpdateLastRestartedSlot

Generated client accounts for UpdateLastRestartedSlot.

UpdateMarketConfig

Generated client accounts for UpdateMarketConfig.

UpdateMarketConfigWithBuffer

Generated client accounts for UpdateMarketConfigWithBuffer.

UpdateOrder

Generated client accounts for UpdateOrder.

UpdatePriceFeedWithChainlink

Generated client accounts for UpdatePriceFeedWithChainlink.

UseClaimableAccount

Generated client accounts for UseClaimableAccount.