gmsol_store::accounts

Struct CreateOrder

pub struct CreateOrder {

    pub owner: Pubkey,
    pub receiver: Pubkey,
    pub store: Pubkey,
    pub market: Pubkey,
    pub user: Pubkey,
    pub order: Pubkey,
    pub position: Option<Pubkey>,
    pub initial_collateral_token: Option<Pubkey>,
    pub final_output_token: Pubkey,
    pub long_token: Option<Pubkey>,
    pub short_token: Option<Pubkey>,
    pub initial_collateral_token_escrow: Option<Pubkey>,
    pub final_output_token_escrow: Option<Pubkey>,
    pub long_token_escrow: Option<Pubkey>,
    pub short_token_escrow: Option<Pubkey>,
    pub initial_collateral_token_source: Option<Pubkey>,
    pub system_program: Pubkey,
    pub token_program: Pubkey,
    pub associated_token_program: Pubkey,

}

Generated client accounts for CreateOrder.

Fields
owner: Pubkey

The owner of the order to be created.

receiver: Pubkey

The receiver of the output funds.

store: Pubkey

Store.

market: Pubkey

Market.

user: Pubkey

User Account.

order: Pubkey

The order to be created.

position: Option<Pubkey>

The related position.

initial_collateral_token: Option<Pubkey>

Initial collateral token / swap in token. Only required by increase and swap orders.

final_output_token: Pubkey

Final output token. Used as collateral token / swap out token for increase and swap orders; and used as final output token for decrease orders.

For the case of increase or swap orders, it will be checked to be a valid collateral / swap out token.

long_token: Option<Pubkey>

Long token of the market.

short_token: Option<Pubkey>

Short token of the market.

initial_collateral_token_escrow: Option<Pubkey>

Initial collateral token escrow account. Only requried by increase and swap orders.

final_output_token_escrow: Option<Pubkey>

Final output token escrow account. Only required by decrease and swap orders.

long_token_escrow: Option<Pubkey>

Long token escrow. Only required by increase and decrease orders.

short_token_escrow: Option<Pubkey>

Short token escrow. Only required by increase and decrease orders.

initial_collateral_token_source: Option<Pubkey>

The source initial token account. Only requried by increase and swap orders.

system_program: Pubkey

The system program.

token_program: Pubkey

The token program.

associated_token_program: Pubkey

The associated token program.