gmsol_store

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

accept_receiver

accept_referral_code

accept_store_authority

auto_deleverage

cancel_order_if_no_position

cancel_referral_code_transfer

check_admin

check_role

claim_fees_from_market

clear_all_prices

close_deposit

close_empty_claimable_account

close_glv_deposit

close_glv_shift

close_glv_withdrawal

close_gt_exchange

close_market_config_buffer

close_order

close_shift

close_withdrawal

confirm_gt_exchange_vault

create_deposit

create_glv_deposit

create_glv_shift

create_glv_withdrawal

create_order

create_shift

create_withdrawal

disable_role

enable_role

execute_decrease_order

execute_deposit

execute_glv_deposit

execute_glv_shift

execute_glv_withdrawal

execute_increase_or_swap_order

execute_shift

execute_withdrawal

get_market_status

get_market_token_price

grant_role

gt_set_exchange_time_window

gt_set_order_fee_discount_factors

gt_set_referral_reward_factors

has_admin

has_role

initialize

initialize_glv

initialize_gt

initialize_market

initialize_market_config_buffer

initialize_market_vault

initialize_oracle

initialize_price_feed

initialize_referral_code

initialize_token_map

insert_address

insert_amount

insert_factor

insert_glv_market

insert_order_fee_discount_for_referred_user

is_token_config_enabled

liquidate

market_transfer_in

migrate_referral_code

prepare_associated_token_account

prepare_gt_exchange_vault

prepare_position

prepare_trade_event_buffer

prepare_user

push_to_market_config_buffer

push_to_token_map

push_to_token_map_synthetic

remove_glv_market

request_gt_exchange

revoke_role

set_expected_provider

set_feed_config

set_market_config_buffer_authority

set_prices_from_price_feed

set_referrer

set_token_map

toggle_feature

toggle_glv_market_flag

toggle_gt_minting

toggle_market

toggle_token_config

token_decimals

token_expected_provider

token_feed

token_name

token_precision

token_timestamp_adjustment

transfer_receiver

transfer_referral_code

transfer_store_authority

update_adl_state

update_glv_config

update_glv_market_config

update_last_restarted_slot

update_market_config

update_market_config_flag

update_market_config_with_buffer

update_order

update_price_feed_with_chainlink

use_claimable_account