gmsol_store::cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

AcceptReceiver

Generated CPI struct of the accounts for AcceptReceiver.

AcceptReferralCode

Generated CPI struct of the accounts for AcceptReferralCode.

AcceptStoreAuthority

Generated CPI struct of the accounts for AcceptStoreAuthority.

CancelOrderIfNoPosition

Generated CPI struct of the accounts for CancelOrderIfNoPosition.

CancelReferralCodeTransfer

Generated CPI struct of the accounts for CancelReferralCodeTransfer.

CheckRole

Generated CPI struct of the accounts for CheckRole.

ClaimFeesFromMarket

Generated CPI struct of the accounts for ClaimFeesFromMarket.

ClearAllPrices

Generated CPI struct of the accounts for ClearAllPrices.

CloseDeposit

Generated CPI struct of the accounts for CloseDeposit.

CloseEmptyClaimableAccount

Generated CPI struct of the accounts for CloseEmptyClaimableAccount.

CloseGlvDeposit

Generated CPI struct of the accounts for CloseGlvDeposit.

CloseGlvShift

Generated CPI struct of the accounts for CloseGlvShift.

CloseGlvWithdrawal

Generated CPI struct of the accounts for CloseGlvWithdrawal.

CloseGtExchange

Generated CPI struct of the accounts for CloseGtExchange.

CloseMarketConfigBuffer

Generated CPI struct of the accounts for CloseMarketConfigBuffer.

CloseOrder

Generated CPI struct of the accounts for CloseOrder.

CloseShift

Generated CPI struct of the accounts for CloseShift.

CloseWithdrawal

Generated CPI struct of the accounts for CloseWithdrawal.

ConfigurateGt

Generated CPI struct of the accounts for ConfigurateGt.

ConfirmGtExchangeVault

Generated CPI struct of the accounts for ConfirmGtExchangeVault.

CreateDeposit

Generated CPI struct of the accounts for CreateDeposit.

CreateGlvDeposit

Generated CPI struct of the accounts for CreateGlvDeposit.

CreateGlvShift

Generated CPI struct of the accounts for CreateGlvShift.

CreateGlvWithdrawal

Generated CPI struct of the accounts for CreateGlvWithdrawal.

CreateOrder

Generated CPI struct of the accounts for CreateOrder.

CreateShift

Generated CPI struct of the accounts for CreateShift.

CreateWithdrawal

Generated CPI struct of the accounts for CreateWithdrawal.

DisableRole

Generated CPI struct of the accounts for DisableRole.

EnableRole

Generated CPI struct of the accounts for EnableRole.

ExecuteDecreaseOrder

Generated CPI struct of the accounts for ExecuteDecreaseOrder.

ExecuteDeposit

Generated CPI struct of the accounts for ExecuteDeposit.

ExecuteGlvDeposit

Generated CPI struct of the accounts for ExecuteGlvDeposit.

ExecuteGlvShift

Generated CPI struct of the accounts for ExecuteGlvShift.

ExecuteGlvWithdrawal

Generated CPI struct of the accounts for ExecuteGlvWithdrawal.

ExecuteIncreaseOrSwapOrder

Generated CPI struct of the accounts for ExecuteIncreaseOrSwapOrder.

ExecuteShift

Generated CPI struct of the accounts for ExecuteShift.

ExecuteWithdrawal

Generated CPI struct of the accounts for ExecuteWithdrawal.

GrantRole

Generated CPI struct of the accounts for GrantRole.

HasRole

Generated CPI struct of the accounts for HasRole.

Initialize

Generated CPI struct of the accounts for Initialize.

InitializeGlv

Generated CPI struct of the accounts for InitializeGlv.

InitializeGt

Generated CPI struct of the accounts for InitializeGt.

InitializeMarket

Generated CPI struct of the accounts for InitializeMarket.

InitializeMarketConfigBuffer

Generated CPI struct of the accounts for InitializeMarketConfigBuffer.

InitializeMarketVault

Generated CPI struct of the accounts for InitializeMarketVault.

InitializeOracle

Generated CPI struct of the accounts for InitializeOracle.

InitializePriceFeed

Generated CPI struct of the accounts for InitializePriceFeed.

InitializeReferralCode

Generated CPI struct of the accounts for InitializeReferralCode.

InitializeTokenMap

Generated CPI struct of the accounts for InitializeTokenMap.

InsertConfig

Generated CPI struct of the accounts for InsertConfig.

InsertGlvMarket

Generated CPI struct of the accounts for InsertGlvMarket.

MarketTransferIn

Generated CPI struct of the accounts for MarketTransferIn.

MigrateReferralCode

Generated CPI struct of the accounts for MigrateReferralCode.

PositionCut

Generated CPI struct of the accounts for PositionCut.

PrepareAssociatedTokenAccount

Generated CPI struct of the accounts for PrepareAssociatedTokenAccount.

PrepareGtExchangeVault

Generated CPI struct of the accounts for PrepareGtExchangeVault.

PreparePosition

Generated CPI struct of the accounts for PreparePosition.

PrepareTradeEventBuffer

Generated CPI struct of the accounts for PrepareTradeEventBuffer.

PrepareUser

Generated CPI struct of the accounts for PrepareUser.

PushToMarketConfigBuffer

Generated CPI struct of the accounts for PushToMarketConfigBuffer.

PushToTokenMap

Generated CPI struct of the accounts for PushToTokenMap.

PushToTokenMapSynthetic

Generated CPI struct of the accounts for PushToTokenMapSynthetic.

ReadMarket

Generated CPI struct of the accounts for ReadMarket.

ReadMarketWithToken

Generated CPI struct of the accounts for ReadMarketWithToken.

ReadTokenMap

Generated CPI struct of the accounts for ReadTokenMap.

RemoveGlvMarket

Generated CPI struct of the accounts for RemoveGlvMarket.

RequestGtExchange

Generated CPI struct of the accounts for RequestGtExchange.

RevokeRole

Generated CPI struct of the accounts for RevokeRole.

SetExpectedProvider

Generated CPI struct of the accounts for SetExpectedProvider.

SetFeedConfig

Generated CPI struct of the accounts for SetFeedConfig.

SetMarketConfigBufferAuthority

Generated CPI struct of the accounts for SetMarketConfigBufferAuthority.

SetPricesFromPriceFeed

Generated CPI struct of the accounts for SetPricesFromPriceFeed.

SetReferrer

Generated CPI struct of the accounts for SetReferrer.

SetTokenMap

Generated CPI struct of the accounts for SetTokenMap.

ToggleFeature

Generated CPI struct of the accounts for ToggleFeature.

ToggleGTMinting

Generated CPI struct of the accounts for ToggleGTMinting.

ToggleMarket

Generated CPI struct of the accounts for ToggleMarket.

ToggleTokenConfig

Generated CPI struct of the accounts for ToggleTokenConfig.

TransferReceiver

Generated CPI struct of the accounts for TransferReceiver.

TransferReferralCode

Generated CPI struct of the accounts for TransferReferralCode.

TransferStoreAuthority

Generated CPI struct of the accounts for TransferStoreAuthority.

UpdateAdlState

Generated CPI struct of the accounts for UpdateAdlState.

UpdateGlvConfig

Generated CPI struct of the accounts for UpdateGlvConfig.

UpdateGlvMarketConfig

Generated CPI struct of the accounts for UpdateGlvMarketConfig.

UpdateLastRestartedSlot

Generated CPI struct of the accounts for UpdateLastRestartedSlot.

UpdateMarketConfig

Generated CPI struct of the accounts for UpdateMarketConfig.

UpdateMarketConfigWithBuffer

Generated CPI struct of the accounts for UpdateMarketConfigWithBuffer.

UpdateOrder

Generated CPI struct of the accounts for UpdateOrder.

UpdatePriceFeedWithChainlink

Generated CPI struct of the accounts for UpdatePriceFeedWithChainlink.

UseClaimableAccount

Generated CPI struct of the accounts for UseClaimableAccount.