gmsol_store::gmsol_store

Function auto_deleverage

pub fn auto_deleverage<'info>(
    ctx: Context<'_, '_, 'info, 'info, PositionCut<'info>>,
    nonce: [u8; 32],
    recent_timestamp: i64,
    size_delta_in_usd: u128,
    execution_fee: u64,
) -> Result<()>

Perform an ADL (Auto-Deleveraging) by keepers.

Accounts

See the documentation for the accounts.

Arguments

nonce: The nonce used to derive the order PDA address.

recent_timestamp: A recent blockchain timestamp for validation.

execution_fee: The execution fee claimed to be used by the keeper.

Errors

The authority must be a signer with the ORDER_KEEPER role.

The owner must be the position owner.

The user must be initialized and corresponding to the owner.

The store must be initialized.

The token_map must be initialized and authorized by the store.

The oracle must be initialized, cleared and store-owned.

The market must be initialized, enabled, store-owned and match the position’s market. The market must be in ADL state.

The order must be uninitialized with address matching PDA from the store, owner, nonce and other expected seeds.

The position must be initialized, owned by the owner and store and eligible for ADL.

The event must be a valid trade event buffer owned by the store and authority.

The long_token and short_token must match those defined in the market.

The long_token_escrow and short_token_escrow must be valid order-owned escrow accounts for their respective tokens.

The long_token_vault and short_token_vault must be valid store-owned market vault accounts for their tokens.

The claimable_long_token_account_for_user must be a store-owned, owner-delegated claimable account for long token.

The claimable_short_token_account_for_user must be a store-owned, owner-delegated claimable account for short token.

The claimable_pnl_token_account_for_holding must be a store-owned, holding-delegated claimable account for PnL token.

Price feed accounts must be valid and provided in the market’s sorted token list order.

The ADL feature must be enabled in the store.

Oracle prices must be valid and complete.

Execution must complete successfully.