gmsol_store::gmsol_store

Function claim_fees_from_market

pub fn claim_fees_from_market(
    ctx: Context<'_, '_, '_, '_, ClaimFeesFromMarket<'_>>,
) -> Result<u64>

Claim fees from the given market.

Accounts

See the documentation for the accounts.

Return

Returns the claimed amount in base units of the token.

Errors

The authority must be a signer and be the designated fee receiver in the given store.

The store must be an initialized Store account owned by this program.

The market must be an initialized Market account owned by this program and associated with the given store.

The token being claimed must be one of the market’s configured collateral tokens.

All provided token accounts must match their expected addresses.

The market must maintain valid balance requirements after the claim.