gmsol_store::gmsol_store

Function close_deposit

pub fn close_deposit<'info>(
    ctx: Context<'_, '_, 'info, 'info, CloseDeposit<'info>>,
    reason: String,
) -> Result<()>

Close a deposit, either by the owner or by keepers.

Accounts

See the documentation for the accounts.

Arguments

reason: The reason for closing the deposit.

Errors

This instruction will fail if:

The executor is not a signer or is neither the deposit owner nor an ORDER_KEEPER in the store.

The store is not properly initialized.

The owner does not match the deposit’s owner.

The provided token mint accounts do not match those recorded in the deposit.

The deposit is not initialized, not owned by the store, or not owned by the specified owner.

Any escrow account is not owned by the deposit or does not match the deposit records.

Any associated token account address is invalid.

The deposit is not in a cancelled or completed state when closed by a non-owner.