gmsol_store::gmsol_store

Function close_withdrawal

pub fn close_withdrawal<'info>(
    ctx: Context<'_, '_, 'info, 'info, CloseWithdrawal<'info>>,
    reason: String,
) -> Result<()>

Close a withdrawal, either by the owner or by keepers.

Accounts

See the documentation for the accounts.

Arguments

reason: The reason for closing the withdrawal.

Errors

This instruction will fail if:

The executor is not a signer or is neither the withdrawal owner nor an ORDER_KEEPER in the store.

The store is not properly initialized.

The owner does not match the withdrawal owner.

The token mint accounts do not match those recorded in the withdrawal.

The withdrawal is not initialized, not owned by the store, or not owned by the specified owner.

Any required escrow accounts are not properly initialized or not owned by the withdrawal.

Any associated token accounts have invalid addresses.

The withdrawal is not in a cancelled or completed state when the executor is not the owner