gmsol_store::gmsol_store

Function create_glv_withdrawal

pub fn create_glv_withdrawal<'info>(
    ctx: Context<'_, '_, 'info, 'info, CreateGlvWithdrawal<'info>>,
    nonce: [u8; 32],
    params: CreateGlvWithdrawalParams,
) -> Result<()>

Create GLV withdrawal.

Accounts

See the documentation for the accounts.

Arguments

nonce: A 32-byte used to derive the address of the GLV withdrawal.

params: The parameters for creating the GLV withdrawal.

Errors

The owner must be a signer and have sufficient balance for the execution fee and rent.

The store must be properly initialized.

The market must be:

Properly initialized

Enabled

Owned by the store

One of the markets in the glv

The glv must be:

Properly initialized

Owned by the store

The glv_withdrawal must be:

Uninitialized

A PDA derived from:

the SEED of GlvWithdrawal

store

owner

nonce

Token requirements:

glv_token must be:

Properly initialized

The GLV token of the glv

market_token must be:

Properly initialized

The market token of the market

All other tokens must be properly initialized

Source requirements:

glv_token_source must be:

Properly initialized

A GLV token account

Have sufficient balance

Have the owner as its authority

Escrow requirements:

Must correspond to their respective tokens

Must be owned by the glv_withdrawal

All token programs must match their corresponding token accounts