gmsol_store::gmsol_store

Function create_order

pub fn create_order<'info>(
    ctx: Context<'_, '_, 'info, 'info, CreateOrder<'info>>,
    nonce: [u8; 32],
    params: CreateOrderParams,
) -> Result<()>

Create an order by the owner.

Accounts

See the documentation for the accounts.

Arguments

nonce: Nonce bytes used to derive the address for the order.

params: Order Parameters specifying the market, order kind, and other details.

Errors

This instruction will fail if:

The owner is not a signer or has insufficient balance for the execution fee and rent.

The store is not properly initialized.

The market is not initialized, is disabled, or not owned by the store.

The user is not initialized or does not correspond to the owner. The address must be a valid PDA derived from the owner and expected seeds.

The order is not uninitialized or the address is not a valid PDA derived from the owner, nonce and expected seeds.

For increase/decrease orders:

The position is missing, not validly initialized, or not owned by both the owner and store.

The long_token or short_token are missing, or do not match the those defined in the market.

The long_token_escrow or short_token_escrow are missing, not valid escrow accounts for long_token or short_token respectively, or not owned by the order.

For increase/swap orders:

The initial_collateral_token is missing or invalid.

The initial_collateral_token_escrow is missing, not a valid escrow account for initial_collateral_token, or not owned by the order.

The initial_collateral_token_source is missing or not a valid source account with owner as the authority.

For decrease/swap orders:

The final_output_token is invalid.

The final_output_token_escrow is missing, not a valid escrow account for final_output_token, or not owned by the order.

The feature for creating this kind of order is not enabled.

The remaining market accounts do not match the swap parameters, not all enabled or owned by the store.