gmsol_store::gmsol_store

Function execute_decrease_order

pub fn execute_decrease_order<'info>(
    ctx: Context<'_, '_, 'info, 'info, ExecuteDecreaseOrder<'info>>,
    recent_timestamp: i64,
    execution_fee: u64,
    throw_on_execution_error: bool,
) -> Result<()>

Execute a decrease order by keepers.

Accounts

See the documentation for the accounts.

Arguments

recent_timestamp: A timestamp that must be within a recent time window.

execution_fee: The execution fee claimed to be used by the keeper.

throw_on_execution_error: If true, throws an error if order execution fails. If false, silently cancels the order on execution failure.

Errors

The authority must be a signer with the ORDER_KEEPER role in the store.

The store must be initialized.

The token_map must be initialized and authorized by the store.

The oracle must be initialized, cleared and owned by the store.

The market must be initialized, enabled and owned by the store.

The owner must be the owner of the order.

The user must be initialized and associated with the owner.

The order must be:

Initialized and owned by both the store and owner

Associated with the provided market

In the pending state

The position must exist and be validly owned by the owner and store. It must match the order.

The event must be a valid trade event buffer owned by both the store and authority.

The tokens must match those recorded in the order.

All escrow accounts must be valid, recorded in the order and owned by the order.

All vault accounts must be valid market vault accounts and owned by the store.

All claimable token accounts must be valid and properly delegated to their owners.

The remaining accounts must be valid. See the documentation for the accounts for more details.

The feature for executing decrease orders must be enabled in the store.

If throw_on_execution_error is true, any execution failure will throw an error.