gmsol_store::gmsol_store

Function execute_deposit

pub fn execute_deposit<'info>(
    ctx: Context<'_, '_, 'info, 'info, ExecuteDeposit<'info>>,
    execution_fee: u64,
    throw_on_execution_error: bool,
) -> Result<()>

Execute a deposit by keepers.

Accounts

See the documentation for the accounts.

Arguments

execution_fee: The execution fee claimed to be used by the keeper.

throw_on_execution_error: If true, throws an error if execution fails. If false, the deposit will be cancelled instead.

Errors

This instruction will fail if:

The authority is not a signer or is not an ORDER_KEEPER in the store.

The store is not properly initialized.

The token_map is not initialized or not authorized by the store.

The oracle is not initialized, cleared and owned by the store.

The market is not initialized, is disabled, not owned by the store, or does not match the market recorded in the deposit.

The deposit is not initialized, not owned by the store, or not in the pending state.

Any token mint accounts do not match those recorded in the deposit.

Any escrow accounts are not properly owned or not recorded in the deposit.

Any vault accounts are not valid market vaults or do not correspond to the initial tokens.

Any feed accounts in the remaining accounts are invalid or do not match the swap parameters.

Any market accounts in the remaining accounts are disabled, not owned by the store, or do not match the swap parameters.

Any oracle prices from the feed accounts are incomplete or invalid.

The execution fails and throw_on_execution_error is set to true.