gmsol_store::gmsol_store

Function execute_glv_shift

pub fn execute_glv_shift<'info>(
    ctx: Context<'_, '_, 'info, 'info, ExecuteGlvShift<'info>>,
    execution_lamports: u64,
    throw_on_execution_error: bool,
) -> Result<()>

Execute GLV shift.

Accounts

See the documentation for the accounts.

Arguments

execution_lamports: The execution fee claimed to be used by the keeper.

throw_on_execution_error: Whether to throw an error if execution fails.

Errors

The authority must be:

A signer

A ORDER_KEEPER in the store

The store must be properly initialized

The token_map must be:

Properly initialized

Authorized by the store

The oracle must be:

Cleared

Owned by the store

The glv must be:

Properly initialized

Owned by the store

The expected GLV of the GLV shift

The from_market must be:

Enabled

Owned by the store

One of the markets in the glv

The to_market must be:

Enabled

Owned by the store

One of the markets in the glv

The glv_shift must be:

Properly initialized

Owned by the store

Token requirements:

from_market_token must be:

The market token of from_market

Recorded in the GLV shift

to_market_token must be:

The market token of to_market

Recorded in the GLV shift

Vault requirements:

from_market_token_glv_vault must be:

The escrow account for from_market_token in the GLV

Owned by the glv

to_market_token_glv_vault must be:

The escrow account for to_market_token in the GLV

Owned by the glv

from_market_token_vault must be:

The market vault for from_market_token

Owned by the store

Token programs must match the tokens and token accounts

The remaining accounts must be valid (see ExecuteGlvShift docs)

Returns error if execution fails and throw_on_execution_error is true