gmsol_store::gmsol_store

Function execute_increase_or_swap_order

pub fn execute_increase_or_swap_order<'info>(
    ctx: Context<'_, '_, 'info, 'info, ExecuteIncreaseOrSwapOrder<'info>>,
    recent_timestamp: i64,
    execution_fee: u64,
    throw_on_execution_error: bool,
) -> Result<()>

Execute an increase/swap order by keepers.

Accounts

See the documentation for the accounts.

Arguments

recent_timestamp: A recent timestamp used for deriving the claimable accounts.

execution_fee: The execution fee claimed to be used the keeper.

throw_on_execution_error: If true, throws an error if order execution fails. If false, silently cancels the order on execution failure.

Errors

The authority must be a signer and have the ORDER_KEEPER role in the store.

The store must be initialized.

The token_map must be initialized and authorized by the store.

The oracle must be initialized, cleared and owned by the store.

The market must be initialized, enabled and owned by the store. It must also be associated with the order.

The owner must be the owner of the order.

The user must be initialized and associated with the owner.

The order must be:

Initialized and owned by both the store and owner

Associated with the provided market

In a pending state

For increase orders:

The initial_collateral_token must be valid.

The position must exist and be owned by the owner and store. It must match the order.

The event must be a valid trade event buffer owned by both the store and authority.

The long_token and short_token must match those defined in the market.

The corresponding token escrow and vault accounts must be valid, recorded in the order and owned by the order.

For swap orders:

The initial_collateral_token must be valid.

The final_output_token must be valid.

The corresponding escrow and vault accounts must be valid, recorded in the order and owned by the order.

The remaining accounts must be valid. See the documentation for the accounts for more details.

The feature for executing this order type must be enabled in the store.

If throw_on_execution_error is true, any execution failure will throw an error