gmsol_store::gmsol_store

Function execute_shift

pub fn execute_shift<'info>(
    ctx: Context<'_, '_, 'info, 'info, ExecuteShift<'info>>,
    execution_lamports: u64,
    throw_on_execution_error: bool,
) -> Result<()>

Execute a shift by keepers.

Accounts

See the documentation for the accounts.

Arguments

execution_lamports: The execution fee in lamports claimed to be used by the keeper.

throw_on_execution_error: Whether to throw an error if the execution fails.

Errors

The authority must be a signer and have the ORDER_KEEPER role in the store.

The store must be initialized.

The token_map must be initialized and authorized by the store.

The oracle must be initialized, cleared and store-owned.

The from_market must be initialized, enabled and store-owned. It must be the from market of the shift.

The to_market must be initialized, enabled and store-owned. It must be the to market of the shift.

The from_market must be shiftable to the to_market.

The shift must be initialized, store-owned and in the pending state.

The from_market_token must be the market token of the from_market.

The to_market_token must be the market token of the to_market.

The from_market_token_escrow must be a valid shift-owned escrow account for the from_market_token and recorded in the shift.

The to_market_token_escrow must be a valid shift-owned escrow account for the to_market_token and recorded in the shift.

The from_market_token_vault must be the market vault for the from_market_token and store-owned.

The feed accounts must be valid and provided in the same order as the unique sorted list of tokens in the from_market and to_market.

The oracle prices from the feed accounts must be complete and valid.

If throw_on_execution_error is true, returns an error if execution fails.