gmsol_store::gmsol_store

Function initialize

pub fn initialize(
    ctx: Context<'_, '_, '_, '_, Initialize<'_>>,
    key: String,
) -> Result<()>

Create a new Store account.

Accounts

See the documentation for the accounts..

Arguments

key: The name of the store, used as a seed to derive the store account’s address. The length must not exceed MAX_LEN.

Errors

The key must be empty unless the multi-store feature is enabled

The payer must be a signer

The authority must be as signer if it is provided.

The receiver must be as signer if it is provided.

The holding must be as signer if it is provided.

The store must not be initialized

The store address must match the PDA derived from the seed of Store and the SHA-256 hash of key