gmsol_store::gmsol_store

Function initialize_glv

pub fn initialize_glv<'info>(
    ctx: Context<'_, '_, 'info, 'info, InitializeGlv<'info>>,
    index: u16,
    length: u16,
) -> Result<()>

Initialize a GLV token and the corresponding GLV account.

Accounts

See the documentation for the accounts.

Arguments

index: The index of the GLV. Used to derive the GLV token address.

length: The number of markets to include in the GLV.

Errors

The authority must be a signer and have MARKET_KEEPER role in the store.

The store must be properly initialized.

The glv_token must be:

Uninitialized

Address must be PDA derived from GLV_TOKEN_SEED, store and index

The glv must be:

Uninitialized

Address must be PDA derived from the SEED of Glv and the address of the glv_token

The remaining required accounts are documented in InitializeGlv.

The length must be:

Greater than 0

Less than or equal to Glv::MAX_ALLOWED_NUMBER_OF_MARKETS