gmsol_store::gmsol_store

Function initialize_market_vault

pub fn initialize_market_vault(
    ctx: Context<'_, '_, '_, '_, InitializeMarketVault<'_>>,
) -> Result<()>

Initialize a new market vault for a specific token.

This instruction creates a new vault account that will be used to store tokens for a market. The vault is a PDA (Program Derived Address) account that can only be controlled by this program.

Accounts

See the documentation for the accounts.

Errors

The authority must be a signer and have MARKET_KEEPER permissions in the store.

The store must be an initialized store account.

The vault must be an uninitialized account and its address must match the PDA derived from the expected seeds.