gmsol_store::gmsol_store

Function initialize_token_map

pub fn initialize_token_map(
    ctx: Context<'_, '_, '_, '_, InitializeTokenMap<'_>>,
) -> Result<()>

Initialize a new token map account with its store set to store.

Anyone can initialize a token map account without any permissions, but after initialization, only addresses authorized by the store can modify this token map (i.e. have the MARKET_KEEPER role).

Accounts

See the documentation for the accounts.

Errors

The payer must be a signer.

The store must be an initialized Store account owned by the store program.

The token_map must be an uninitialized account.