gmsol_store::gmsol_store

Function insert_address

pub fn insert_address(
    ctx: Context<'_, '_, '_, '_, InsertConfig<'_>>,
    key: String,
    address: Pubkey,
) -> Result<()>

Insert an address value into the store’s global configuration.

This instruction allows a CONFIG_KEEPER to set or update an address value in the store’s configuration. The key must be one of the predefined address keys.

Accounts

See the documentation for the accounts..

Arguments

key: The configuration key to update. Must be a valid address key defined in AddressKey.

address: The address value to store for this configuration key.

Errors

The authority must be a signer and have the CONFIG_KEEPER role in the store.

The provided key must be defined in AddressKey.

The store must be initialized and owned by this program.