gmsol_store::gmsol_store

Function liquidate

pub fn liquidate<'info>(
    ctx: Context<'_, '_, 'info, 'info, PositionCut<'info>>,
    nonce: [u8; 32],
    recent_timestamp: i64,
    execution_fee: u64,
) -> Result<()>

Perform a liquidation by keepers.

Accounts

See the documentation for the accounts.

Arguments

nonce: The nonce used to derive the order PDA address.

recent_timestamp: A recent timestamp that must be within the valid time window.

execution_fee: The execution fee claimed to be used by the keeper.

Errors

The authority must be a signer with the ORDER_KEEPER role in the store.

The owner must be the owner of the position being liquidated.

The user must be an initialized user account corresponding to the owner.

The store must be initialized.

The token_map must be initialized and authorized by the store.

The oracle must be initialized, cleared and owned by the store.

The market must be:

Initialized and enabled

Owned by the store

Associated with the position being liquidated

The order must be:

Uninitialized

Have an address matching the PDA derived from the store, owner, provided nonce and other expected seeds

The position must be:

Validly initialized

Owned by both the owner and store

In a liquidatable state

The event must be a valid trade event buffer owned by both the store and authority.

The long_token and short_token must match those defined in the market.

Token escrow accounts must be:

Valid for their respective tokens

Owned by the order

Market vault accounts must be:

Valid market vault accounts for their respective tokens

Owned by the store

Claimable token accounts must be:

Valid for their respective tokens

Owned by the store

Properly delegated to their owners

Price feed accounts must be:

Valid and complete

Provided in order matching the market’s sorted token list

The liquidation feature must be enabled in the store.

Oracle prices must be valid and complete.