gmsol_store::gmsol_store

Function prepare_position

pub fn prepare_position(
    ctx: Context<'_, '_, '_, '_, PreparePosition<'_>>,
    params: CreateOrderParams,
) -> Result<()>

Prepare the position account for orders.

Accounts

See the documentation for the accounts.

Arguments

params: Order Parameters.

Errors

This instruction will fail if:

The owner is not a signer or has insufficient balance for the rent.

The store is not properly initialized.

The market is not initialized, is disabled, or not owned by the store.

The position address is not a valid PDA derived from the owner and expected seeds.

The position account is neither uninitialized nor validly initialized with store as the store and owner as the owner.