gmsol_store::gmsol_store

Function push_to_market_config_buffer

pub fn push_to_market_config_buffer(
    ctx: Context<'_, '_, '_, '_, PushToMarketConfigBuffer<'_>>,
    new_configs: Vec<EntryArgs>,
) -> Result<()>

Push config items to the given market config buffer account.

This instruction allows the authority to add new configuration items to their market config buffer account. The buffer will be reallocated to accommodate the new items, with the authority paying for any additional rent.

Accounts

See the documentation for the accounts.

Arguments

new_configs: The list of new config items to append to the buffer. Each item consists of a string key and a factor value.

Errors

The authority must be a signer and the owner of the buffer account.

The buffer must be an initialized market config buffer account.

The authority must have enough SOL to pay for any additional rent needed.

The keys in new_configs must be valid MarketConfigKey.