gmsol_store::gmsol_store

Function set_feed_config

pub fn set_feed_config(
    ctx: Context<'_, '_, '_, '_, SetFeedConfig<'_>>,
    token: Pubkey,
    provider: u8,
    feed: Pubkey,
    timestamp_adjustment: u32,
) -> Result<()>

Set the feed config of the given provider for the given token.

Accounts

See the documentation for the accounts.

Arguments

token: The token whose config will be updated.

provider: The index of the provider whose feed config will be updated. Must be a valid PriceProviderKind value.

feed: The new feed address.

timestamp_adjustment: The new timestamp adjustment in seconds.

Errors

The authority must be a signer and a MARKET_KEEPER in the given store.

The store must be an initialized Store account owned by the store program.

The token_map must be an initialized token map account owned by the store.

The given token must exist in the token map.

The provider index must correspond to a valid PriceProviderKind.