gmsol_store::gmsol_store

Function toggle_glv_market_flag

pub fn toggle_glv_market_flag(
    ctx: Context<'_, '_, '_, '_, UpdateGlvMarketConfig<'_>>,
    flag: String,
    enable: bool,
) -> Result<()>

Toggle the given flag of a market in the given GLV.

Accounts

See the documentation for the accounts.

Arguments

flag: The flag to toggle.

enable: The value to toggle to.

Errors

The authority must be:

A signer

Have MARKET_KEEPER role in the store

The store must be properly initialized.

The glv must be:

Properly initialized

Owned by the store

Have the market token in its list of market tokens

The market_token must be:

Properly initialized

Owned by the store

flag must be defined in GlvMarketFlag.