gmsol_store::gmsol_store

Function toggle_token_config

pub fn toggle_token_config(
    ctx: Context<'_, '_, '_, '_, ToggleTokenConfig<'_>>,
    token: Pubkey,
    enable: bool,
) -> Result<()>

Enable or disable the config for the given token.

Accounts

See the documentation for the accounts.

Arguments

token: The token whose config will be updated.

enable: Enable or disable the config.

Errors

The authority must be a signer and a MARKET_KEEPER in the given store.

The store must be an initialized Store account owned by the store program .

The token_map must be an initialized token map account owned by the store.

The given token must exist in the token map.