gmsol_store::gmsol_store

Function transfer_store_authority

pub fn transfer_store_authority(
    ctx: Context<'_, '_, '_, '_, TransferStoreAuthority<'_>>,
) -> Result<()>

Request to transfer the authority (admin) of the given store to a new address.

Note

This instruction only sets next_authority. Use accept_store_authority to complete the transfer.

Accounts

See the documentation for the accounts..

Errors

The authority must be a signer and the current admin of the store.

The store must be an initialized store account owned by the store program.

The next_authority cannot be the same as current next_authority.