gmsol_store::gmsol_store

Function update_adl_state

pub fn update_adl_state<'info>(
    ctx: Context<'_, '_, 'info, 'info, UpdateAdlState<'info>>,
    is_long: bool,
) -> Result<()>

Update the ADL (Auto-Deleveraging) state for the market.

Accounts

See the documentation for the accounts.

Arguments

is_long: Whether to update the ADL state for the long (true) or short (false) side of the market.

Errors

The authority must be a signer and have the ORDER_KEEPER role in the store.

The store must be an initialized Store account owned by the store program.

The oracle must be an initialized Oracle account that is owned by the store.

The market must be enabled and owned by the store.

Price feed accounts must be valid and provided in the market’s sorted token list order.