gmsol_store::gmsol_store

Function update_price_feed_with_chainlink

pub fn update_price_feed_with_chainlink(
    ctx: Context<'_, '_, '_, '_, UpdatePriceFeedWithChainlink<'_>>,
    compressed_report: Vec<u8>,
) -> Result<()>

Updates the price data in a custom price feed account using a signed price report from Chainlink Data Streams. The price feed must be configured to use the Chainlink Data Streams provider.

Accounts

See the documentation for the accounts.

Arguments

signed_report: A signed price report from Chainlink Data Streams containing the price data.

Errors

The authority must be a signer and have the PRICE_KEEPER role in the store.

The store must be an initialized store account

The verifier_account must be a valid Chainlink verifier account.

The price_feed must be initialized, owned by the store, and authorized for the authority.

The chainlink program ID must be trusted in the definition of the ChainlinkDataStreamsInterface.

The price feed must be configured to use ChainlinkDataStreams as its provider.

The signed_report must be:

Decodable as a valid Chainlink price report

Verifiable by the Chainlink Verifier Program

Contain valid data for creating a PriceFeedPrice

The current slot and timestamp must be >= the feed’s last update.

The price data timestamp must be >= the feed’s last price timestamp

The price data must meet all validity requirements (see the update method of PriceFeed).