gmsol_store::instructions

Module roles

Instructions for roles management.

Structs

CheckRole

The accounts definition for check_admin and check_role.

CheckRoleBumps

DisableRole

The accounts definition for disable_role.

DisableRoleBumps

EnableRole

The accounts definition for enable_role.

EnableRoleBumps

GrantRole

The accounts definition for grant_role.

GrantRoleBumps

HasRole

The accounts definition for has_admin and has_role.

HasRoleBumps

RevokeRole

The accounts definition for revoke_role.

RevokeRoleBumps

Functions

has_admin

Verify that the user is an admin of the given store without signing.

has_role

Verify that the authority has the given role in the given store without signing.

unchecked_disable_role

Disable the given role in the data store.